+++ src/browser.ts
@@ -170,6 +170,36 @@ export async function pollForFirstContentfulPaint(driver: webdriver.WebDriver):
 }
 
 /**
+ * Poll for the `window.tachometerResult` global and return it once it is set.
+ * Polls every 50 milliseconds, and returns undefined if no result was found
+ * after 10 seconds. Throws if a value was found, but it was not a number, or it
+ * was a negative number.
+ */
+export async function pollForGlobalResult(driver: webdriver.WebDriver):
+    Promise<number|undefined> {
+  // Both here and for FCP above, we could automatically tune the poll time
+  // after we get our first result, so that when the script is fast we spend
+  // less time waiting, and so that when the script is slow we interfere it
+  // less frequently.
+  for (let waited = 0; waited <= 10000; waited += 50) {
+    await wait(50);
+    const result = await driver.executeScript(
+                       'return window.tachometerResult;') as unknown;
+    if (result !== undefined && result !== null) {
+      if (typeof result !== 'number') {
+        throw new Error(
+            `window.tachometerResult was type ` +
+            `${typeof result}, expected number.`);
+      }
+      if (result < 0) {
+        throw new Error(`window.tachometerResult was negative: ${result}`);
+      }
+      return result;
+    }
+  }
+}
+
+/**
  * https://developer.mozilla.org/en-US/docs/Web/API/PerformanceEntry
  *
  * Note a more complete interface for this is defined in the standard